Resolve issue where DesiredFramesPerSecond would get set to 1 in the Designer
This was due to some drawing code setting it to 1 in case of a renderer
error, which is apparently occurring when the component is displayed in the
Designer. Since this property change affects the saved file, the change is
now avoided when in the Designer and a plain background is drawn instead.