Avoid updating the tree view unnecessarily when using Make Default
Always refreshing the tree when PropertyGridManager.RefreshUI was called seems like a bad idea. This even causes the tree view to get refreshed just because the selected element changes, because that also refreshes the property grid.

For now, the tree view is only explicitly refreshed by "Make Default" when it detects the "Parent" variable of an instance changed.

Refreshing the tree view unnecessarily isn't really a performance concern at the moment, but it is annoying because it destroys the expanded state of the tree nodes. This is of course something that would be better to fix properly, because it should be possible for the tree to get updated without destroying that state.