Avoid exception while handling UnauthorizedAccessException
At least in German locale, the file name is quoted using " instead of '
characters which caused it to fail getting the file name from the
exception message.

In general, parsing the file name from the exception message can be
expected to fail, so I changed the handling to just show the exception
as-is when the file name could not be determined.

The correct fix may be to introduce our own exception class in which we
can pass on the file that was being written explicitly.