Changed the mouse handling for the panning for smoother updating
Previously, the panning was handled before each repaint of the view, which
happen at a fixed 30 fps. Now, the panning is done in response to the
mouse events and this in turn invalidates the view.

This allows the view to update immediately in response to mouse panning,
giving a much more responsive experience.