Removed the SaveProject call in LoadProject
It's in my opinion quite a bad idea to automatically save the whole
project when it gets loaded. Two examples where this goes wrong:

* Version control systems like Perforce require files to be checked out
  before being written to. In effect, this behavior causes all files in
  the project to be checked out just by opening Gum.

* Sometimes I've managed to start Gum without our plugin loaded. This
  would then inadvertedly cause the Standard elements to get written to
  disk, including the set of properties that I'm suppressing in the
  plugin.

In general, I think it's just unexpected behavior. It's also unnecessary
because whatever changes are being made to the project during load, you
can expect Gum to make those same changes again next time it loads the
project.

Finally, if you really want a project to be saved in its entirety, then
there is already a menu item available for this.