fix: a couple binding hiccups (#817)
* Fixes a couple binding hiccups

Ensures previous source getter/setter are cleared when detaching from previous bindingcontext.

Allows binding to readonly source properties by only attempting to build a setter and getter, and gracefully handles cases where a setter is not available.

Adds a test case to verify that binding to a readonly source property updates the target.

* oops forgot to uncomment these