Update sonarscanner.yml (#1489)
Removing the cache, which should force the install of the dotnet-scanner to be the latest always.  Since this runs AFTER The merge into the primary branch, it shouldn't impact other PR's.

We can always add it back after the version gets "bumped"

Also, this fixes #1353 but technically we were not using the github specific action, so we were not impacted by this.  However, just in case we were somehow impacted by the changes I've removed the cache so we get the newest scanner.